backend/requirements.txt
fastapi
uvicorn
langchain
langgraph
langchain-openai
sqlalchemy
aiosqlite
python-dotenv
requests
beautifulsoup4
feedparser
pillow
pytesseract
tavily-python
pydantic-settings
werkzeug
scikit-learn
opencv-python-headless
numpy
python-multipart
